fix circular reference
