+++ assets/model_monitoring/components/src/action_analyzer/contracts/action_sample.py
@@ -3,7 +3,35 @@
 
 """Action sample class."""
 
-from action_analyzer.contracts.action import ActionSample
+
+class ActionSample:
+    """Action sample class."""
+
+    def __init__(self,
+                 question: str,
+                 answer: str,
+                 debugging_info: str,
+                 prompt_flow_input: str) -> None:
+        """Create an action sample.
+
+        Args:
+            question(str): the input question of the flow.
+            answer(str): the output answer of the flow.
+            debugging_info(str): the json string of debugging info in a span tree structure.
+            prompt_flow_input(str): the json str of prompt flow input.
+        """
+        self.question = question
+        self.answer = answer
+        self.debugging_info = debugging_info
+        self.prompt_flow_input = prompt_flow_input
+
+    def to_json_str(self) -> str:
+        """Convert an action sample object to json string."""
+        attribute_dict = self.__dict__
+        json_out = {}
+        for key, val in attribute_dict.items():
+            json_out[convert_to_camel_case(key)] = val
+        return json.dumps(json_out, default=np_encoder)
 
 
 class IndexActionSample(ActionSample):
