+++ assets/model_monitoring/components/src/action_analyzer/contracts/actions/low_retreival_score_index_action.py
@@ -3,8 +3,8 @@
 
 """LowRetreivalScoreIndexAction Class."""
 
-from action_analyzer.contracts.action import ActionType, Action
-from action_analyzer.contracts.index_action_contracts import IndexActionSample
+from action_analyzer.contracts.actions.action import ActionType, Action
+from action_analyzer.contracts.action_sample import IndexActionSample
 from shared_utilities.constants import (
     ACTION_DESCRIPTION
 )
