+++ assets/model_monitoring/components/src/action_analyzer/contracts/actions/metrics_violation_index_action.py
@@ -3,8 +3,8 @@
 
 """MetricsViolationIndexAction Class."""
 
-from action_analyzer.contracts.action import ActionType, Action
-from action_analyzer.contracts.index_action_contracts import IndexActionSample
+from action_analyzer.contracts.actions.action import ActionType, Action
+from action_analyzer.contracts.action_sample import IndexActionSample
 from shared_utilities.constants import (
     ACTION_DESCRIPTION
 )
